Basic Instruction datatype and parser.
I still don't really know how I'll implement the loop. What I
basically want is it to be a simple Jump. I'd like to avoid having the
nested instruction list like in most naive implementations.
For the moment its argument is just a label. A Jmp and an End have a
matching label.